refactor some of the tasks out of the mojo into separate task classes
