+++ pom.xml
@@ -70,7 +70,7 @@
             <plugin>
                 <groupId>org.codehaus.plexus</groupId>
                 <artifactId>plexus-component-metadata</artifactId>
-                <version>1.5.5</version>
+                <version>1.7.1</version>
                 <executions>
                     <execution>
                         <goals>
